Fix 182 security vulnerabilities, logic bugs, and technical issues
Web Backend (53 files):
- CRITICAL: Import security.py in settings (was dead code: Argon2, CSP, JWT validation)
- CRITICAL: Add 2FA check in login flow (was completely bypassed)
- CRITICAL: Verify OAuth ID tokens with provider JWKS (Google/Apple)
- CRITICAL: Fix mass assignment in sync _create_object (field allowlist)
- CRITICAL: Fix XXE in OFX importer (use defusedxml)
- CRITICAL: Fix IDOR in transaction serializer (ownership validation)
- CRITICAL: Fix OAuth state leaked in banking response
- CRITICAL: Fix unvalidated redirect_uri in banking OAuth
- HIGH: Fix ALLOWED_HOSTS accepting empty string in production
- HIGH: Move RateLimitMiddleware after AuthenticationMiddleware
- HIGH: Fix CORS allow-all with credentials in development
- HIGH: Restrict UserViewSet to read/update only (was full CRUD)
- HIGH: Validate refresh token ownership in logout
- HIGH: Invalidate all tokens on password change
- HIGH: Fix banking provider init order (super().__init__ before attrs)
- HIGH: Fix Stripe webhook method names and service signatures
- HIGH: Fix budget signals referencing non-existent models/fields
- HIGH: Fix AI priority type mismatch (string vs integer)
- HIGH: Fix cross-tenant data access in AI batch tasks
- HIGH: Remove @csrf_exempt from AI ChatView
- HIGH: Fix IP spoofing via X-Forwarded-For (use TRUSTED_PROXY_COUNT)
- HIGH: Fix lru_cache on middleware instance methods
- HIGH: Fix SSRF in profile picture fetch (URL validation + size limits)
- MEDIUM: Fix naive datetime in tokens (use timezone.now())
- MEDIUM: Fix email_verified field name in token invalidation
- MEDIUM: Use PBKDF2 instead of unsalted SHA-256 for backup codes
- MEDIUM: Make TOTP replay prevention atomic (select_for_update)
- MEDIUM: Handle IntegrityError in TOTP setup_device
- MEDIUM: Fix non-atomic budget item updates (transaction.atomic)
- MEDIUM: Restrict ChatMessageSerializer role to user/assistant only
- MEDIUM: Fix O(n^2) anomaly detection to O(n) using dict
- MEDIUM: Fix AI insights view wrong field names
- MEDIUM: Fix recommendations type/priority mismatches
- MEDIUM: Replace MD5 with SHA256 in categorization cache
- MEDIUM: Add field allowlist in sync serialization
- MEDIUM: Validate sync conflict data with full_clean()
- MEDIUM: Add exception exclusion in Celery retry logic
- MEDIUM: Require auth for API docs/schema endpoints
- MEDIUM: Fix _raw_delete to use proper .delete()
- MEDIUM: Fix sanitization regex false positives
- MEDIUM: Reduce subscription cache TTL
- MEDIUM: Fix cascading profile save signal
- LOW: Fix duplicate SIMPLE_JWT definitions (security.py overrides)
- LOW: Fix non-atomic email rate limiting

Infrastructure (15 files):
- CRITICAL: Fix shell code injection via DB_PASSWORD in entrypoint.sh
- CRITICAL: Enable Redis authentication (requirepass)
- CRITICAL: Disable dangerous Redis commands (FLUSHALL, CONFIG, DEBUG)
- CRITICAL: Fix XSS via |safe filter in transaction templates
- CRITICAL: Fix XSS in messages template (remove |safe)
- CRITICAL: Fix default Flower credentials in docker-compose.prod
- HIGH: Remove direct Django port exposure in docker-compose
- HIGH: Add non-root user in dev Dockerfile
- HIGH: Bind dev database/Redis to 127.0.0.1
- HIGH: Remove || true from CI security checks
- HIGH: Fix CSP unsafe-eval in nginx
- HIGH: Add IP restriction for admin panel
- HIGH: Add SRI to CDN scripts in base template
- MEDIUM: Fix deploy heredoc variable expansion
- MEDIUM: Pin GitHub Actions versions

Desktop App (17 files):
- CRITICAL: Fix Zip Slip vulnerability in backup restore
- CRITICAL: Encrypt token fallback file storage
- HIGH: Change default API URL to HTTPS
- HIGH: Add SSL disable warning and production guard
- HIGH: Mask database URL in logs
- HIGH: Fix unbound variable in backup restore
- HIGH: Add allowlist for notification preference setattr
- HIGH: Add allowlist for User.from_dict setattr
- HIGH: Mask PII in auth logs
- MEDIUM: Replace MD5 with SHA256 in import dedup
- MEDIUM: Fix float to Decimal for financial amounts
- MEDIUM: Fix connectivity check (use API server not Google DNS)
- MEDIUM: Add allowlist for SyncRecord.from_dict setattr
- MEDIUM: Fix XOR cipher fallback warning
- MEDIUM: Validate URLs before opening in browser
- LOW: Rename ImportError to DataImportError (avoid shadowing)
- LOW: Remove email from User.__repr__

Mobile App (15 files):
- CRITICAL: Apply certificate pinning to OkHttpClient (was never applied)
- CRITICAL: Fix cleartext traffic in network security config
- CRITICAL: Add authentication gate in MainActivity
- CRITICAL: Validate deep links require auth
- CRITICAL: Fix SQL injection in database prepopulation
- HIGH: Disable android:allowBackup for financial app
- HIGH: Use fallbackToDestructiveMigrationOnDowngrade
- HIGH: Implement token refresh in AuthInterceptor
- HIGH: Fix backup rules (exclude tokens, fix db name)
- MEDIUM: Use BigDecimal for financial amounts in DTOs
- MEDIUM: Safe enum parsing with fallback in BudgetDto
- MEDIUM: Disable retryOnConnectionFailure for mutations
- MEDIUM: Fix Gson date format UTC handling
- MEDIUM: Override toString() to mask tokens in UserPreferences
- MEDIUM: Add ProGuard rules to strip debug logs
- LOW: Replace android.util.Log with Timber in DatabaseCallback

+++ pecunia-mobile/data/remote/dto/TransactionDto.kt
@@ -2,7 +2,9 @@ package com.pecunia.data.remote.dto
 
 import com.pecunia.data.local.database.entities.TransactionEntity
 import com.pecunia.data.local.database.entities.TransactionType
+import com.google.gson.annotations.JsonAdapter
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 /**
  * Data Transfer Object for Transaction from API.
@@ -15,7 +17,7 @@ data class TransactionDto(
     val userId: String,
 
     @SerializedName("amount")
-    val amount: Double,
+    val amount: BigDecimal,
 
     @SerializedName("currency")
     val currency: String,
@@ -63,7 +65,7 @@ data class TransactionDto(
         return TransactionEntity(
             id = id,
             userId = userId,
-            amount = amount,
+            amount = amount.toDouble(),
             currency = currency,
             type = TransactionType.valueOf(type.uppercase()),
             category = category,
@@ -87,7 +89,7 @@ data class TransactionDto(
  */
 data class TransactionRequest(
     @SerializedName("amount")
-    val amount: Double,
+    val amount: BigDecimal,
 
     @SerializedName("currency")
     val currency: String = "USD",
@@ -128,7 +130,7 @@ data class TransactionRequest(
          */
         fun fromEntity(entity: TransactionEntity): TransactionRequest {
             return TransactionRequest(
-                amount = entity.amount,
+                amount = BigDecimal.valueOf(entity.amount),
                 currency = entity.currency,
                 type = entity.type.name,
                 category = entity.category,
@@ -181,13 +183,13 @@ data class TransactionResponse(
  */
 data class TransactionSummaryResponse(
     @SerializedName("total_income")
-    val totalIncome: Double,
+    val totalIncome: BigDecimal,
 
     @SerializedName("total_expense")
-    val totalExpense: Double,
+    val totalExpense: BigDecimal,
 
     @SerializedName("net_balance")
-    val netBalance: Double,
+    val netBalance: BigDecimal,
 
     @SerializedName("category_breakdown")
     val categoryBreakdown: List<CategoryBreakdown>,
@@ -207,10 +209,10 @@ data class CategoryBreakdown(
     val category: String,
 
     @SerializedName("amount")
-    val amount: Double,
+    val amount: BigDecimal,
 
     @SerializedName("percentage")
-    val percentage: Double,
+    val percentage: BigDecimal,
 
     @SerializedName("transaction_count")
     val transactionCount: Int
